combined cl and wnb and made a new category for airplanes
