+++ app/src/main/AndroidManifest.xml
@@ -105,7 +105,6 @@ Redistribution and use in source and binary forms, with or without modification,
         <activity android:name=".PfdActivity"/>
         <activity android:name=".RegisterActivity"/>
         <activity android:name=".ThreeDActivity"/>
-        <activity android:name=".WnbActivity"/>
         <activity
             android:name="com.ds.avare.MainActivity"
             android:launchMode="singleTask"
